Update README
Sorry to make a pull request like this. But I didnt saw anyway to pull a issue point here. (I am not used to with Github yet)
I was cloning this kernel source to build a custom nethunter kernel for my Redmi 8
It was not working with Git Clone then I found out there is some file with uppercase and lowercase file with same name that might causing the issue after downloading zipfile.
Please let me know if this is the issue that might be causing this bug or if there is something I am missing. 
(Also if possible do tell me what should I do if I see similar issue like this on github again 😅 and let the repositor owner know the issue with the files)
And thanks a lot for your contribution, Time and everything and reading the my issue